rename nldns to nodelocaldns
