+++ plugin/pkg/shoot/nodelocaldns/nodelocaldns_suite_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package nldns_test
+package nodelocaldns_test
 
 import (
 	"testing"
